% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootjack.R
\name{bootjack}
\alias{bootjack}
\title{Bootstrap-jacknife of flow calibration statistics}
\usage{
bootjack(
  flows,
  stats = "nse",
  nSample = 1000,
  waterYearMonth = 10,
  startYear = NULL,
  endYear = NULL,
  minYears = 10
)
}
\arguments{
\item{flows}{Required. Data frame of date, observed and simulated flows.}

\item{stats}{Required. Vector of goodness of fit statistics to be used.}

\item{waterYearMonth}{Required. Month of beginning of water year. Default is \code{10}
(October). If the calendar year is required, set \code{waterYearMonth = 13}.}

\item{startYear}{Optional. First year of data to be used. If \code{NULL} then not used.}

\item{endYear}{Optional. Last year of data to be used. If \code{NULL} then not used.}

\item{minYears}{Required. Minimum number years to be used. Default is 10.}

\item{nSampls}{Required. Number of samples for bootstrapping.}
}
\value{
Returns a list containing 1) \code{statsBoot} (array of bootstrap stats), and
2) \code{statsJack} (array of jackknife stats).
}
\description{
Bootstrap-jacknife of flow calibration statistics
}
\examples{
\dontrun{
NSE_stats <- bootjack(flows)}
}
\seealso{
\code{\link{read_hcdn}}
}
\author{
Kevin Shook
}
