% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conus_hcdn_bootjack.R
\name{conus_hcdn_bootjack}
\alias{conus_hcdn_bootjack}
\title{Jackkife after bootstrap for all HCDN sites}
\usage{
conus_hcdn_bootjack(
  hcdn_sites = NULL,
  NetCDF_file = NULL,
  sim_var = "kge",
  GOF_stat = c("NSE", "KGE"),
  nSample = 1000,
  waterYearMonth = 10,
  startYear = NULL,
  endYear = NULL,
  minDays = 100,
  minYears = 10
)
}
\arguments{
\item{hcdn_sites}{Required. Data frame of HCDN sites. Must contain a field called \option{hcdn_site"}. The data frame
\code{hcdn_conus_sites} will work. You can subset this data frame if you want to use fewer sites.}

\item{NetCDF_file}{Required. NetCDF file containing modelled and gauged flows for CONUS.}

\item{sim_var}{Required. Name of variable containing simulated flows in \code{NetCDF}.}

\item{GOF_stat}{Required. Name(s) of simulation goodness of fit statistic(s) to be calculated. Currently both \code{NSE} and \code{KGE} are supported.}

\item{nSample}{Required. Number of samples for bootstrapping.}

\item{waterYearMonth}{Required. Month of beginning of water year. Default is \code{10}
(October). If the calendar year is required, set \code{waterYearMonth = 13}.}

\item{startYear}{Optional. First year of data to be used. If \code{NULL} then not used.}

\item{endYear}{Optional. Last year of data to be used. If \code{NULL} then not used.}

\item{minDays}{Required. Minimum number of days per year with valid (i.e. greater than 0) flows. Default is 100.}

\item{minYears}{Required. Minimum number years to be used. Default is 10.}
}
\value{
Returns a data frame containing the following variables:
\itemize{
 \item{\code{hcdn_site}}{HCDN site number}
 \item{\code{lat}}{HCDN site latitude}
 \item{\code{lon}}{HCDN site longitude}
 \item{\code{GOF_stat}}{Goodness of fit statistics (i.e. NSE or KGE)}
 \item{\code{seJack}}{standard error of jacknife}
 \item{\code{seBoot}}{standard error of bootstrap}
 \item{\code{p05, p50, p95}}{the 5th, 50th and 95th percentiles of the estimates}
 \item{\code{score}}{the jackknife score}
 \item{\code{biasJack}}{the bias of the jackknife}
 \item{\code{biasBoot}}{the bias of the bootstrap}
 \item{\code{seJab}}{the standard error of the jackknife after bootstrap}
}


}
\description{
Jackkife after bootstrap for all HCDN sites
}
\examples{
\dontrun{
hcdn <- conus_hcdn_bootjack(hcdn_sites = sites, NetCDF_file = "results_hcdn_flow.nc")
}

}
