% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows_1030500.R
\docType{data}
\name{flows_1030500}
\alias{flows_1030500}
\title{Observed and simulated flows for a single location}
\format{
A data frame with 6940 rows and 3 variables:
\describe{
  \item{date}{Date of flows}
  \item{obs}{observed flows (m\eqn{^3}{^3})/s)}
  \item{sim}{simulated flows (m\eqn{^3}{^3})/s)}
}
}
\usage{
flows_1030500
}
\description{
A data frame containing observed and simulated flows for USGS site 1030500
}
\keyword{datasets}
