% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_estimate_uncertainties.R
\name{ggplot_estimate_uncertainties}
\alias{ggplot_estimate_uncertainties}
\title{Plots uncertainties in model error estimates}
\usage{
ggplot_estimate_uncertainties(JAB_stats, fill_colour = "orange")
}
\arguments{
\item{JAB_stats}{Required. Data frame of jackknife after boot statistics for a large number
of model runs, as produced by \code{conus_hcdn_bootjack}.}

\item{fill_colour}{Required. Color used to fill the plot of 2 x the Jackknife estimate of the standard error.}
}
\value{
Returns a \code{ggplot2} object of the plots, faceted by goodness of fit statistic, i.e. NSE/KGE.
The confidence interval (difference between the 95^th^ and 5^th^ quantiles, and the value of
2 x the Bootstrap estimate of the standard error are plotted as lines. The values of
2 x the Jackknife estimate of the standard error are plotted as filled)
}
\description{
Plots uncertainties in model error estimates
}
\examples{
\dontrun{ p <- ggplot_estimate_uncertainties(all_stats, "orange")
}
}
\seealso{
\code{\link{conus_hcdn_bootjack}}
}
\author{
Kevin Shook
}
