% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hcdn.R
\name{read_hcdn}
\alias{read_hcdn}
\title{Reads simulated and observed values from HCDN netcdf file for a single location}
\usage{
read_hcdn(nc_file, hcdn, obsName = "obs", simName = "sig1")
}
\arguments{
\item{nc_file}{Required. netCDF file to read from.}

\item{hcdn}{Required. Location to extract data.}

\item{obsName}{Required. Name for variable containing observations. Default is "obs".}

\item{simName}{Required. Name for variable containing simulations. Default is "sig1".}
}
\value{
Returns a data frame containing the date, observed and simulated flows. The name of the
observed flow variable is \code{obs}, the name of the simulated flow variable is \code{sim}.
}
\description{
Reads simulated and observed values from HCDN netcdf file for a single location
}
\examples{
\dontrun{
flows <- read_hcdn(nc_file = "results_hcdn_flow.nc", hcdn = 1030500)
}
}
