% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumboot-package.R
\docType{package}
\name{gumboot-package}
\alias{gumboot-package}
\title{Bootstrap Analyses of Hydrological Model Error}
\description{
Does jackknife after bootstrap analyses of the error in hydrological models by
estimating the empirical probability distributions of NSE (Nash-Sutcliffe efficiency) and KGE (Kling-Gupta efficiency)
estimators.
}
\section{Funding}{


The package was partly funded by the Global institute for Water Security (GIWS; \href{https://water.usask.ca/}{https://water.usask.ca/}) and the Global Water Futures (GWF; \href{https://gwf.usask.ca/}{https://gwf.usask.ca/}) program.
}

\references{
The package code is described in: \cr
 \cite{Clark et al. (2021), "The abuse of popular performance metrics in hydrologic modeling", Water Resources Research, <doi:10.1029/2020WR029001>.}
}
\author{
\strong{Coded by:} Martyn Clark and Kevin Shook

\strong{Maintained by:} Kevin Shook \email{kevin.shook@usask.ca}
}
